add: products and categories stat cards
